backend/requirements.txt
flask
flask-cors
pypdf
faiss-cpu
numpy
scikit-learn
requests
gunicorn
groq